scripts/requirements.txt
pyyaml==6.0.1
requests==2.31.0
pandas==2.2.2
numpy==1.26.4
pytz==2024.1
scikit-learn
joblib
openai==2.12.0